+++ src/inmanta/compiler/__init__.py
@@ -18,11 +18,22 @@
 import logging
 import os
 import sys
+from itertools import chain
 from typing import Dict, List, Optional, Set
 
+import inmanta.ast.type as inmanta_type
 import inmanta.execute.dataflow as dataflow
 from inmanta import const
-from inmanta.ast import AttributeException, CompilerException, DoubleSetException, LocatableString, MultiException, Range
+from inmanta.ast import (
+    AttributeException,
+    CompilerException,
+    DoubleSetException,
+    LocatableString,
+    MultiException,
+    Namespace,
+    Range,
+)
+from inmanta.ast.entity import Entity
 from inmanta.ast.statements.define import DefineEntity, DefineRelation, PluginStatement
 from inmanta.compiler import config as compiler_config
 from inmanta.execute import scheduler
@@ -46,17 +57,37 @@ def do_compile(refs={}):
     LOGGER.debug("Starting compile")
 
     (statements, blocks) = compiler.compile()
-    sched = scheduler.Scheduler(compiler_config.datatrace_enable.get())
+    sched = scheduler.Scheduler(compiler_config.track_dataflow())
     try:
         success = sched.run(compiler, statements, blocks)
     except CompilerException as e:
+        if compiler_config.dataflow_graphic_enable.get():
+            show_dataflow_graphic(sched, compiler)
         compiler.handle_exception(e)
 
     LOGGER.debug("Compile done")
 
     if not success:
         sys.stderr.write("Unable to execute all statements.\n")
+    if compiler_config.dataflow_graphic_enable.get():
+        show_dataflow_graphic(sched, compiler)
     return (sched.get_types(), compiler.get_ns())
+
+
+def show_dataflow_graphic(scheduler, compiler):
+    from inmanta.execute.dataflow.graphic import GraphicRenderer
+
+    types: Dict[str, inmanta_type.Type] = scheduler.get_types()
+    ns: Namespace = compiler.get_ns()
+    config_ns: Namespace = ns.get_child("__config__")
+    GraphicRenderer.view(
+        config_ns.get_scope().slots.values(),
+        list(
+            chain.from_iterable(
+                tp.get_all_instances() for tp in types.values() if isinstance(tp, Entity) and tp.namespace is config_ns
+            )
+        ),
+    )
 
 
 def anchormap(refs={}):