+++ src/inmanta/server/server.py
@@ -213,7 +213,8 @@ angular.module('inmantaApi.config', []).constant('inmantaConfig', {
     @protocol.handle(methods_v2.get_api_docs)
     async def get_api_docs(self, format: Optional[str]) -> ReturnValue[Union[OpenAPI, str]]:
         url_map = self._server._transport.get_global_url_map(self._server.get_slices().values())
-        openapi = OpenApiConverter(url_map)
+        feature_manager = self.feature_manager
+        openapi = OpenApiConverter(url_map, feature_manager)
         # Get rid of none values with custom json encoder
         openapi_json_str = openapi.generate_openapi_json()
         if format == "openapi":