+++ CHANGELOG.md
@@ -16,9 +16,11 @@
  - Use asyncio subprocess instead of tornado subprocess (#1792)
  - Added warning for incorrect database migration script names (#1912)
  - Agent manager remains consistent when the database connection is lost (#1893)
+ - Ensure correct version is used in api docs (#1994)
+ - Fixed double assignment error resulting from combining constructor kwargs with default values (#2003)
 
 ## Added
- - Experimental data trace and root cause application (#1820, #1831, #1821)
+ - Experimental data trace, root cause and graphic data flow visualization applications (#1820, #1831, #1821, #1822)
  - Warning when shadowing variable (#1366, #1918)
  - Added support for compiler warnings (#1779, #1905, #1906)
  - Added support for DISABLED flag for database migration scripts (#1913)
@@ -28,6 +30,8 @@
  - Added support to pause an agent (#1128)
  - Added --no-tag option to module tool (#1939)
  - Added base exception for plugins and corresponding documentation (#1205)
+ - Added tags to openapi definition (#1751)
+ - Added support to pause an agent (#1128, #1982)
 
 # v 2020.1 (2020-02-19) Changes in this release: